+++ bayespy/inference/vmp/nodes/tests/test_beta.py
@@ -9,9 +9,6 @@
 Unit tests for `beta` module.
 """
 
-import warnings
-warnings.simplefilter("error")
-
 import numpy as np
 from scipy import special