+++ bayespy/inference/vmp/nodes/tests/test_gaussian_markov_chain.py
@@ -9,9 +9,6 @@
 Unit tests for gaussian_markov_chain module.
 """
 
-import warnings
-warnings.simplefilter("error")
-
 import numpy as np
 
 from ..gaussian_markov_chain import GaussianMarkovChain