+++ bayespy/inference/vmp/nodes/tests/test_mixture.py
@@ -10,7 +10,6 @@ Unit tests for mixture module.
 """
 
 import warnings
-warnings.simplefilter("error")
 
 import numpy as np